Extracts BackstagePasses into a separate class
